Split CLI build integration tests
